Pessoas em um raio melhorado, create adicionado

+++ GeoDengue/src/br/edu/ufcg/geodengue/shared/RaioDTO.java
@@ -6,19 +6,19 @@ public class RaioDTO implements Serializable {
 
 	private static final long serialVersionUID = 7789787012069960980L;
 
-	private int raio;
+	private double raio;
 	private double latitude;
 	private double longitude;
 	
 	public RaioDTO() { }
 	
-	public RaioDTO(double latitude, double longitude, int raio) {
+	public RaioDTO(double latitude, double longitude, double raio) {
 		this.raio = raio;
 		this.latitude = latitude;
 		this.longitude = longitude;
 	}
 
-	public int getRaio() {
+	public double getRaio() {
 		return raio;
 	}
